requirements.txt
discord.py==1.6.0
colorama==0.4.4
python-dotenv==0.15.0